restconf: establish-subscription RPC for subscribed notifications
This is a very bare-bones implementation of the subscribed
notifications concept from RFC 8639 [1]. We only support the
establish-subscription RPC to the NETCONF event stream.

This patch *does not* ship the code which implements the endpoint for
client connection because this patch is already too large and the
endpoint is not strictly necessary for the RPC to work. It will come in
the follow-up patch.

[1] https://datatracker.ietf.org/doc/rfc8639/

Change-Id: I0217e5abc56cfd73859dbcc610fb1f342dc33a10
Signed-off-by: Mattias Walström <lazzer@gmail.com>