restconf: terminate subsc. notification streams after 60s with no client
We allow anonymous users to create notification streams. However, such
users will not have permissions to delete the subscriptions which opens
a potential for resource exhaustion. To prevent this, we now terminate
the notification streams after 60 seconds of inactivity.

The period is configurable via the Server constructor, but the default
is 60 seconds.

Change-Id: Id399d108069704fa8ab79b9ef6ab855dd62cfc8d
Signed-off-by: Mattias Walström <lazzer@gmail.com>