port to libyang v4
- the prefix of the "ietf-netconf-with-defaults" module has changed,
- RPC parsing needs a flag to reject unknown input data nodes,
- due to the "printed context" mess, we now need an extra flag to ensure
  that we can still access the YANG schema source.

The last item is interesting. We could also read directly from the
on-disk location, but that looks a bit meh. It's also a bit annoying
that one has to change this thing over a lot of places, but I think that
this is much nicer than trying to force a change over an existing
connection (e.g., via a connection-specific call in the Server::Server
ctor which might hit some internal locks depending on how the sysrepo
state look like).

I was considering adding a wrapper for `ly_ctx_get_options()` and a
check in Server::Server, but then I decided that I won't bother.

Oh, and this "disable printed context" is also needed for the RESTCONF
*client* which is used in the unit test. The `lyd_parse_op` cannot parse
notifications *iff* the `notifications` module from RFC 5277 is missing
*and* when the libyang context does not have the "parsed" info from the
`ietf-yang-types` available. The printed context, as provided by
sysrepo, does not have that data available. This is all papered over as
long as the context flag setting works, which is the case with
post-4.2.4 version of sysrepo (still untagged). Yay.

We still set these flags to a single value in all of the unit tests and
in the main() as well to maintain the remaining bits of our sanity, of
course.

Change-Id: I142f650affb53411b5c1dd81311400a2aa6dec20
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8990
Bug: https://github.com/CESNET/libyang/pull/2448
Depends-on: https://github.com/sysrepo/sysrepo/commit/6dc5641762962b93d54a54443e2fd43aa319a7a6
Signed-off-by: Mattias Walström <lazzer@gmail.com>